refactor(mcp): ToolResult ヘルパを shared に抽出し MarkdownCodec を generic 化
Task 7 の legal tools が使う本文パイプラインの共有下地。tools.test.ts は
無変更のまま緑(挙動保存の証明)。
